Moved repository trait in another folder.
