+++ src/Repositories/Seat/Filters/NamedIdFilter.php
@@ -1,0 +1,58 @@
+<?php
+
+/**
+ * MIT License.
+ *
+ * Copyright (c) 2019. Felix Huber
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+namespace Seat\Services\Repositories\Seat;
+
+use Illuminate\Support\Collection;
+use Seat\Eveapi\Models\Character\CharacterInfo;
+use Seat\Eveapi\Models\Universe\UniverseName;
+
+/**
+ * Class DataTablesFilter.
+ * @package Seat\Services\Repositories\Seat
+ */
+trait NamedIdFilter
+{
+    /**
+     * @param $keyword
+     *
+     * @return \Illuminate\Support\Collection
+     */
+    public function getIdsForNames($keyword) : Collection
+    {
+        return UniverseName::where('name', 'like', '%' . $keyword . '%')
+            ->get()
+            ->map(function ($resolved_id) {
+                return $resolved_id->entity_id;
+            })
+            ->merge(CharacterInfo::where('name', 'like', '%' . $keyword . '%')
+                ->get()
+                ->map(function ($character_info) {
+                    return $character_info->character_id;
+                })
+            );
+    }
+}
